unified sun with other towers

+++ src/game/effects/Effect.java
@@ -25,8 +25,8 @@ abstract public class Effect
 		ageInSeconds += deltaTime;
 		
 		// Sets new position of Effect
-		posX += velocityX*deltaTime;
-		posY += velocityY*deltaTime;
+		posX += (int) (velocityX*deltaTime);
+		posY += (int) (velocityY*deltaTime);
 		
 		// see if stardust hit enemy
 		List<Enemy> enemies = game.enemies;
